// src/modules/user/enums/user.enum.ts
export enum ENUM_USER_SIGN_UP_FROM {
    DASHBOARD = 'DASHBOARD',
    APP = 'APP',
    SEED = 'SEED',
}

export enum ENUM_USER_STATUS {
    PENDING = 'PENDING',
    ACTIVE = 'ACTIVE',
    INACTIVE = 'INACTIVE',
    DELETED = 'DELETED',
    BLOCKED = 'BLOCKED',
}

export enum ENUM_USER_GENDER {
    MALE = 'MALE',
    FEMALE = 'FEMALE',
}